try to change connections num in page provider

+++ backend/scalestore/rdma/MessageHandler.cpp
@@ -168,7 +168,7 @@ void MessageHandler::startThread() {
                  rdma::InitMessage* initMsg = (rdma::InitMessage*)cm.getGlobalBuffer().allocate(sizeof(rdma::InitMessage));
                  // fill init messages
                  initMsg->mbOffset = 0;  // No MB offset
-                 initMsg->plOffset = mhEndpoint.request;
+                 initMsg->plOffset = (uintptr_t)mhEndpoint.request;
                  initMsg->bmId = nodeId;
                  initMsg->type = rdma::MESSAGE_TYPE::Init;
                  cm.exchangeInitialMesssage(*(mhEndpoint.rctx), initMsg);
